requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.1.1
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
